% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_year.R
\name{academic_year}
\alias{academic_year}
\alias{is_academic_year}
\alias{as_academic_year}
\title{Utilities for working with academic years}
\usage{
academic_year(
  x = integer(),
  boundary = getOption("acadyr.academic_year_boundary", as.Date("2020-09-01"))
)

is_academic_year(x)

as_academic_year(
  x,
  boundary = getOption("acadyr.academic_year_boundary", as.Date("2020-09-01")),
  ...
)
}
\arguments{
\item{x}{\itemize{
\item For \code{academic_year()}: A numeric vector
\item For \code{as_academic_year()}: An object to coerce
\item For \code{is_academic_year()}: An object to test
}}

\item{boundary}{A boundary date where the new academic year should begin.
Note that only the month/day of this date is used. This can be set using
the global option \code{acadyr.academic_year_boundary}. E.g. to use academic
years starting on August 1, use the following:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(acadyr.academic_year_boundary = as.Date("2020-08-01"))
}\if{html}{\out{</div>}}}

\item{...}{Passed on to methods
\itemize{
\item For \code{academic_year()}: A \verb{<financial_year>} vector
\item For \code{as_academic_year()}: A \verb{<financial_year>} vector
\item For \code{is_academic_year()}: A \verb{<logical>} vector
}}
}
\description{
Utilities for working with academic years
}
\details{
Note that many countries use a slightly different academic year start date
each year, meaning that most use-cases for these utilities will necessitate
some approximation.
}
\examples{
# Numbers are treated as if they represent years
academic_year(2005:2015)

# Dates are mapped to the year they fall within
dates <- as.Date(c("2020-08-01", "2020-09-01"))
academic_year(dates)

# The year they fall within depends on the year boundary, which
# defaults to September 01
academic_year(dates, boundary = as.Date("2020-08-01"))

# as_academic_year() also understands character vectors:
as_academic_year(c("2020/21", "2021/22"))
}
