% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acadyr-package.R
\docType{package}
\name{acadyr-package}
\alias{acadyr}
\alias{acadyr-package}
\title{acadyr: Convenience Functions for Working With Non-Calendar Years in R}
\description{
Provides S3 classes for working with academic and financial year data types.
}
\author{
\strong{Maintainer}: Jacob Scott \email{jscott2718@gmail.com}

}
\keyword{internal}
