% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_year.R
\name{scale_x_academic_year}
\alias{scale_x_academic_year}
\alias{scale_y_academic_year}
\title{Position scales for academic years}
\usage{
scale_x_academic_year(labels = ~format(academic_year(.)), ...)

scale_y_academic_year(labels = ~format(academic_year(.)), ...)
}
\arguments{
\item{labels, ...}{Passed to \verb{ggplot2::scale_(x|y)_continuous()}}
}
\description{
Position scales for academic years
}
