% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_year.R
\name{scale_x_financial_year}
\alias{scale_x_financial_year}
\alias{scale_y_financial_year}
\title{Position scales for financial years}
\usage{
scale_x_financial_year(labels = ~format(financial_year(.)), ...)

scale_y_financial_year(labels = ~format(financial_year(.)), ...)
}
\arguments{
\item{labels, ...}{Passed to \verb{ggplot2::scale_(x|y)_continuous()}}
}
\description{
Position scales for financial years
}
