% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year-utils.R
\name{set_year_boundary}
\alias{set_year_boundary}
\alias{year_boundary<-}
\title{Set the boundary when a new year begins}
\usage{
set_year_boundary(x, value)

year_boundary(x) <- value
}
\arguments{
\item{x}{An academic/financial year object}

\item{value}{A date}
}
\description{
Set the boundary when a new year begins
}
